moved function to common
