+++ lib/workers/repository/process/lookup/common.ts
@@ -1,0 +1,14 @@
+import type { LookupUpdateConfig } from './types';
+
+export function mergeConfigConstraints(
+  config: LookupUpdateConfig
+): LookupUpdateConfig {
+  if (config?.extractedConstraints) {
+    config.constraints = {
+      ...config.extractedConstraints,
+      ...config.constraints,
+    };
+    delete config.extractedConstraints;
+  }
+  return config;
+}
