+++ lib/workers/repository/process/lookup/index.spec.ts
@@ -16,7 +16,6 @@ import { id as pep440VersioningId } from '../../../../modules/versioning/pep440'
 import { id as poetryVersioningId } from '../../../../modules/versioning/poetry';
 import type { LookupUpdateConfig } from './types';
 import * as lookup from '.';
-import { mergeConfigConstraints } from '.';
 
 jest.mock('../../../../modules/datasource/docker');
 
@@ -1720,59 +1719,6 @@ describe('workers/repository/process/lookup/index', () => {
           updateType: `rollback`,
         },
       ]);
-    });
-
-    it('overrides extracted config with user config', () => {
-      const config: LookupUpdateConfig = {
-        datasource: '',
-        depName: '',
-        versioning: '',
-        rangeStrategy: 'pin',
-      };
-      config.constraints = {
-        constraint1: 'configValue1',
-        constraint2: 'configValue2',
-        constraint3: 'configValue3',
-      };
-      config.extractedConstraints = {
-        constraint3: 'extractedValue3',
-        constraint4: 'exractedValue4',
-      };
-      expect(mergeConfigConstraints(config)).toMatchObject({
-        datasource: '',
-        depName: '',
-        versioning: '',
-        rangeStrategy: 'pin',
-        constraints: {
-          constraint1: 'configValue1',
-          constraint2: 'configValue2',
-          constraint3: 'configValue3',
-          constraint4: 'exractedValue4',
-        },
-      });
-    });
-
-    it('sets config with extracted config', () => {
-      const config: LookupUpdateConfig = {
-        datasource: '',
-        depName: '',
-        versioning: '',
-        rangeStrategy: 'pin',
-      };
-      config.extractedConstraints = {
-        constraint3: 'extractedValue3',
-        constraint4: 'exractedValue4',
-      };
-      expect(mergeConfigConstraints(config)).toMatchObject({
-        datasource: '',
-        depName: '',
-        versioning: '',
-        rangeStrategy: 'pin',
-        constraints: {
-          constraint3: 'extractedValue3',
-          constraint4: 'exractedValue4',
-        },
-      });
     });
   });
 });
